JobApi/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
distlib==0.3.0
Django==3.0.6
django-cors-headers==3.5.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.21
